#8638 exasol dialogs moved from ext.exasol to ext/exasol.ui, configurators-part moved from managers in ext.exasol
