+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/manager/ExasolForeignKeyManager.java
@@ -3,10 +3,7 @@ package org.jkiss.dbeaver.ext.exasol.manager;
 import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolTable;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolTableForeignKey;
-import org.jkiss.dbeaver.ext.exasol.model.ExasolTableForeignKeyColumn;
-import org.jkiss.dbeaver.ext.exasol.model.ExasolTableUniqueKey;
 import org.jkiss.dbeaver.ext.exasol.tools.ExasolUtils;
-import org.jkiss.dbeaver.ext.exasol.ui.ExasolCreateForeignKeyDialog;
 import org.jkiss.dbeaver.model.DBPEvaluationContext;
 import org.jkiss.dbeaver.model.DBUtils;
 import org.jkiss.dbeaver.model.edit.DBECommandContext;
@@ -19,9 +16,7 @@ import org.jkiss.dbeaver.model.messages.ModelMessages;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.dbeaver.model.struct.DBSObject;
 import org.jkiss.dbeaver.model.struct.cache.DBSObjectCache;
-import org.jkiss.dbeaver.ui.UITask;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -48,40 +43,7 @@ public class ExasolForeignKeyManager
             "FK"
         );
         foreignKey.setName(getNewConstraintName(monitor, foreignKey));
-
-        return new UITask<ExasolTableForeignKey>() {
-            @Override
-            protected ExasolTableForeignKey runTask() {
-                ExasolCreateForeignKeyDialog editPage = new ExasolCreateForeignKeyDialog("Create Foreign Key", foreignKey);
-
-                if (!editPage.edit()) {
-                    return null;
-                }
-                foreignKey.setName(editPage.getName());
-                foreignKey.setReferencedConstraint((ExasolTableUniqueKey)editPage.getUniqueConstraint());
-                foreignKey.setEnabled(editPage.isEnabled());
-
-                List<ExasolTableForeignKeyColumn> columns = new ArrayList<>();
-                int cnt = 0;
-                for (ExasolCreateForeignKeyDialog.FKColumnInfo column : editPage.getColumns()) {
-                    try {
-                        ExasolTable refTable = foreignKey.getReferencedConstraint().getTable();
-                        columns.add(new ExasolTableForeignKeyColumn(
-                            foreignKey,
-                            table.getAttribute(monitor, column.getOwnColumn().getName()),
-                            refTable.getAttribute(monitor, column.getRefColumn().getName()),
-                            ++cnt));
-                    } catch (DBException e) {
-                        log.error("Could not get Attribute Information from Table");
-                        return null;
-                    }
-                }
-
-                foreignKey.setColumns(columns);
-
-                return foreignKey;
-            }
-        }.execute();
+        return foreignKey;
     }
 
     @Override
