+++ plugins/org.jkiss.dbeaver.ext.exasol/src/org/jkiss/dbeaver/ext/exasol/manager/ExasolUserManager.java
@@ -4,11 +4,12 @@ import org.eclipse.jface.dialogs.IDialogConstants;
 import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.ext.exasol.ExasolConstants;
 import org.jkiss.dbeaver.ext.exasol.ExasolMessages;
+import org.jkiss.dbeaver.ext.exasol.ExasolUserType;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolDataSource;
 import org.jkiss.dbeaver.ext.exasol.model.ExasolPriority;
 import org.jkiss.dbeaver.ext.exasol.model.security.ExasolUser;
 import org.jkiss.dbeaver.ext.exasol.tools.ExasolUtils;
-import org.jkiss.dbeaver.ext.exasol.ui.ExasolUserDialog;
+//import org.jkiss.dbeaver.ext.exasol.ui.ExasolUserDialog;
 import org.jkiss.dbeaver.ext.exasol.ui.ExasolUserQueryPassword;
 import org.jkiss.dbeaver.model.DBPDataSource;
 import org.jkiss.dbeaver.model.DBUtils;
@@ -47,7 +48,8 @@ public class ExasolUserManager extends SQLObjectEditor<ExasolUser, ExasolDataSou
     protected ExasolUser createDatabaseObject(DBRProgressMonitor monitor,
                                               DBECommandContext context, Object container, Object copyFrom, Map<String, Object> options)
         throws DBException {
-        return new UITask<ExasolUser>() {
+    	return new ExasolUser((ExasolDataSource) container, "user", "", "", "password", "", ExasolUserType.LOCAL);
+        /*return new UITask<ExasolUser>() {
             @Override
             protected ExasolUser runTask() {
                 ExasolUserDialog dialog = new ExasolUserDialog(UIUtils.getActiveWorkbenchShell(), (ExasolDataSource) container);
@@ -64,7 +66,7 @@ public class ExasolUserManager extends SQLObjectEditor<ExasolUser, ExasolDataSou
                     dialog.getUserType());
                 return user;
             }
-        }.execute();
+        }.execute();*/
     }
 
     @Override
